Enemies Scripting: Moving Enemies, Enemies that Shoot, Enemies that Stand still and face the player, and the Boss (The Philanthropist)
Setup most of the basic information and base for creating the main enemy types: Flying, Grounded, Shooting, Moving, Stationary, Player-Conscious, and Bosses.
This also includes the base for "Thresholds" the player activates by crossing them. Mostly used for locking off boss areas and cutscenes.
Bullets are also included and can be customized and specialized.
The player can heal, but atm it's only the additional health and stamina => The attack doesn't cost stamina yet, but will on the next update
Finding an alternative to fixedupdate has been a failure so most of the things involving collisions will now take place there, for the time being. This has been a hit to performance but atm there doesn't seem to be a way to do it without using Rigidbody. => If this becomes a bigger product the switch from CC to Rigidbody will need to be looked out.